Refactor graph data structures and algorithms for improved consistency and functionality; update edge and node manipulation methods to enhance clarity and maintainability

+++ include/typedefs.h
@@ -10,7 +10,7 @@
 
 
 
-typedef unsigned int INDEX;
+typedef size_t INDEX;
 typedef INDEX Label;
 typedef INDEX NodeId;
 typedef std::vector<Label> Labels;
